+++ loqusdb/utils/load.py
@@ -32,6 +32,7 @@ def load_database(
     skip_case_id=False,
     gq_threshold=None,
     snv_gq_only=False,
+    keep_chr_prefix=False,
     qual_gq=False,
     case_id=None,
     max_window=3000,
@@ -51,6 +52,7 @@ def load_database(
           family_type(str): Format of family file
           skip_case_id(bool): If no case information should be added to variants
           gq_threshold(int): If only quality variants should be considered
+          keep_chr_prefix(bool): Retain chr/CHR/Chr prefix when present
           qual_gq(bool): Use QUAL field instead of GQ format tag to gate quality
           case_id(str): If different case id than the one in family file should be used
           max_window(int): Specify the max size for sv windows
@@ -68,7 +70,7 @@ def load_database(
     nr_variants = None
     vcf_individuals = None
     if variant_file:
-        vcf_info = check_vcf(variant_file)
+        vcf_info = check_vcf(variant_file, keep_chr_prefix)
         nr_variants = vcf_info["nr_variants"]
         variant_type = vcf_info["variant_type"]
         vcf_files.append(variant_file)
@@ -78,7 +80,7 @@ def load_database(
     nr_sv_variants = None
     sv_individuals = None
     if sv_file:
-        vcf_info = check_vcf(sv_file, "sv")
+        vcf_info = check_vcf(sv_file, keep_chr_prefix, "sv")
         nr_sv_variants = vcf_info["nr_variants"]
         vcf_files.append(sv_file)
         sv_individuals = vcf_info["individuals"]
@@ -86,7 +88,7 @@ def load_database(
     profiles = None
     matches = None
     if profile_file:
-        profiles = get_profiles(adapter, profile_file)
+        profiles = get_profiles(adapter, profile_file, keep_chr_prefix)
         ###Check if any profile already exists
         matches = profile_match(
             adapter, profiles, hard_threshold=hard_threshold, soft_threshold=soft_threshold
@@ -152,6 +154,7 @@ def load_database(
                 skip_case_id=skip_case_id,
                 gq_threshold=gq_threshold if not snv_gq_only or variant_type == "snv" else None,
                 qual_gq=qual_gq,
+                keep_chr_prefix=keep_chr_prefix,
                 max_window=max_window,
                 variant_type=variant_type,
                 genome_build=genome_build,
@@ -200,6 +203,7 @@ def load_variants(
     skip_case_id=False,
     gq_threshold=None,
     qual_gq=False,
+    keep_chr_prefix=False,
     max_window=3000,
     variant_type="snv",
     genome_build=None,
@@ -213,6 +217,7 @@ def load_variants(
         case_obj(Case): dict with case information
         skip_case_id (bool): whether to include the case id on variant level
                              or not
+        keep_chr_prefix(bool): Retain chr/CHR/Chr prefix when present
         gq_threshold(int)
         qual_gq(bool): whether to use QUAL instead of GQ
         max_window(int): Specify the max size for sv windows
@@ -242,6 +247,7 @@ def load_variants(
                 case_id,
                 gq_threshold,
                 qual_gq,
+                keep_chr_prefix,
                 ignore_gq_if_unset,
                 genome_build=genome_build,
             )
@@ -263,7 +269,7 @@ def load_variants(
     return nr_inserted
 
 
-def load_profile_variants(adapter, variant_file):
+def load_profile_variants(adapter, variant_file, keep_chr_prefix=None):
     """
 
     Loads variants used for profiling
@@ -275,7 +281,7 @@ def load_profile_variants(adapter, variant_file):
 
     """
 
-    vcf_info = check_vcf(variant_file)
+    vcf_info = check_vcf(variant_file, keep_chr_prefix)
     variant_type = vcf_info["variant_type"]
 
     if variant_type != "snv":
@@ -284,5 +290,5 @@ def load_profile_variants(adapter, variant_file):
 
     vcf = get_vcf(variant_file)
 
-    profile_variants = [build_profile_variant(variant) for variant in vcf]
+    profile_variants = [build_profile_variant(variant, keep_chr_prefix) for variant in vcf]
     adapter.add_profile_variants(profile_variants)